refactor: app folder in src refactor: all configs in .ts
