+++ jest.setup.ts
@@ -1,9 +1,9 @@
 // Optional: configure or set up a testing framework before each test.
-// If you delete this file, remove `setupFilesAfterEnv` from `jest.config.js`
+// If you delete this file, remove `setupFilesAfterEnv` from `jest.config.*`
 
 // Used for __tests__/testing-library.js
 // Learn more: https://github.com/testing-library/jest-dom
-import '@testing-library/jest-dom/extend-expect';
+import '@testing-library/jest-dom';
 
 // To get 'next/router' working with tests
 jest.mock('next/router', () => require('next-router-mock'));
