--- next.config.js
@@ -1,7 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-  // reactStrictMode: false,
-};
-
-module.exports = nextConfig;
