+++ package.json
@@ -23,7 +23,6 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "build:static": "next build && next export",
     "format": "prettier ./ --write",
     "lint": "next lint",
     "start": "next start",
@@ -48,16 +47,17 @@
     "@testing-library/jest-dom": "latest",
     "@testing-library/react": "latest",
     "@testing-library/user-event": "latest",
+    "@types/jest": "latest",
     "@types/node": "latest",
     "@types/react": "latest",
-    "babel-jest": "latest",
+    "@types/react-dom": "latest",
     "eslint": "latest",
     "eslint-config-next": "latest",
-    "eslint-plugin-testing-library": "latest",
     "jest": "latest",
     "jest-environment-jsdom": "latest",
     "next-router-mock": "latest",
     "prettier": "latest",
-    "typescript": "latest"
+    "ts-node": "latest",
+    "typescript": "^5"
   }
 }
