+++ src/app/home/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata, NextPage } from 'next';
 import { Stack, Typography } from '@mui/material';
-import { AppLink } from 'src/components';
+import { AppLink } from '@/components';
 import DemoAppAlert from '../dev/components/DemoAppAlerts';
 import DemoAppButton from '../dev/components/DemoAppButton';
 import DemoAppIcon from '../dev/components/DemoAppIcon';
