+++ src/components/AppIconButton/AppIconButton.test.tsx
@@ -1,8 +1,8 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import AppIconButton, { MUI_ICON_BUTTON_COLORS } from './AppIconButton';
 import { APP_ICON_SIZE } from '../config';
-import { ObjectPropByName, capitalize, randomColor, randomText } from '../../utils';
-import { ICONS } from '../AppIcon/AppIcon';
+import { capitalize, randomColor, randomText } from '../../utils';
+import { ICONS } from '../AppIcon/config';
 
 const ComponentToTest = AppIconButton;
 
@@ -12,10 +12,10 @@ function randomPropertyName(obj: object): string {
   return propertyName;
 }
 
-function randomPropertyValue(obj: object): unknown {
-  const propertyName = randomPropertyName(obj);
-  return (obj as ObjectPropByName)[propertyName];
-}
+// function randomPropertyValue(obj: object): unknown {
+//   const propertyName = randomPropertyName(obj);
+//   return (obj as ObjectPropByName)[propertyName];
+// }
 
 /**
  * Tests for <AppIconButton/> component
